Removed second activity just used for selecting feeds
