+++ app/src/main/res/layout/opml_selection.xml
@@ -1,39 +1,44 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
 
     <Button
-            android:id="@+id/butConfirm"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentBottom="true"
-            android:layout_alignParentRight="true"
-            android:layout_alignParentEnd="true"
-            style="@style/Widget.MaterialComponents.Button.TextButton"
-            android:layout_margin="8dp"
-            android:text="@string/confirm_label"/>
+        android:id="@+id/butConfirm"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentBottom="true"
+        android:layout_alignParentRight="true"
+        android:layout_alignParentEnd="true"
+        android:layout_margin="8dp"
+        android:text="@string/confirm_label"
+        style="@style/Widget.MaterialComponents.Button.TextButton" />
 
     <Button
-            android:id="@+id/butCancel"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentBottom="true"
-            android:layout_toLeftOf="@+id/butConfirm"
-            android:layout_toStartOf="@+id/butConfirm"
-            style="@style/Widget.MaterialComponents.Button.TextButton"
-            android:layout_margin="8dp"
-            android:text="@string/cancel_label"/>
+        android:id="@+id/butCancel"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignParentBottom="true"
+        android:layout_toLeftOf="@+id/butConfirm"
+        android:layout_toStartOf="@+id/butConfirm"
+        android:layout_margin="8dp"
+        android:text="@string/cancel_label"
+        style="@style/Widget.MaterialComponents.Button.TextButton" />
 
     <ListView
-            android:id="@+id/feedlist"
-            android:layout_width="match_parent"
-            android:layout_height="0dp"
-            android:layout_above="@id/butConfirm"
-            android:layout_alignParentTop="true"
-            tools:listitem="@android:layout/simple_list_item_multiple_choice">
-    </ListView>
+        android:id="@+id/feedlist"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:layout_above="@id/butConfirm"
+        android:layout_alignParentTop="true"
+        tools:listitem="@android:layout/simple_list_item_multiple_choice" />
+
+    <ProgressBar
+        android:id="@+id/progressBar"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_centerInParent="true" />
 
 </RelativeLayout>
